Worked!!
This version is the best till now. The validity equal to 100 filter conditions works as expected. All other columns can be changed without affecting the logic, which is what we need to work in the Google Sheets as a Database, allowing us to apply the changes to previous registers without worrying about overwriting from the collector.

The name of the Spreadsheet is still hardcoded and for now it does not need to be changed